pyproject.toml
[tool.poetry]
name = "tradesignals-platform"
version = "0.1.0"
description = "Tradesignals Platform application server and data pipelines."
authors = ["Mac Anderson <devops@tradesignals.io>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://tradesignals.io"
repository = "https://github.com/Tradesignals-io/tradesignals-platform"
documentation = "https://docs.tradesignals.io"

packages = [
    { include = "api", from = "src" },
    { include = "pipelines", from = "src" },
    { include = "core", from = "src" },
    { include = "cli", from = "src" },
    { include = "streaming", from = "src" },
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Quantitative Finance",
    "Intended Audience :: Data Scientists",
    "Intended Audience :: Financial Analysts",
    "Intended Audience :: Traders",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Financial",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Office/Business :: Financial :: Trading",
    "Topic :: Office/Business :: Financial :: Analysis",
]

[tool.poetry.urls]
Changelog = "https://github.com/Tradesignals-io/tradesignals-platform/releases"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
dotenv-derive = "^0.1.0"
confluent-kafka = {extras = ["avro"], version = "^2.6.1"}
python-dotenv = "^1.0.1"
attrs = "^24.2.0"
cattrs = "^24.1.2"
pandas-market-calendars = "^4.4.2"
pandas = "^2.2.3"
uvloop = "^0.21.0"
requests = "^2.32.3"
tornado = "^6.4.2"
click = ">=8.1.7"
databento = "^0.46.0"
orjson = "^3.10.12"
tradesignals-python = "^0.1.0a10"
boto3 = "^1.35.78"
nox = "^2024.10.9"
scipy = "^1.14.1"
numpy = "^2.2.0"
scikit-learn = "^1.6.0"
websocket-client = "^1.8.0"
rel = "^0.4.9.19"

[tool.poetry.dev-dependencies]
Pygments = ">=2.18.0"
black = ">=24.10.0"
coverage = {extras = ["toml"], version = ">=7.6.9"}
darglint = ">=1.8.1"
flake8 = ">=7.1.1"
flake8-bandit = ">=4.1.1"
flake8-bugbear = ">=24.10.31"
flake8-docstrings = ">=1.7.0"
flake8-rst-docstrings = ">=0.3.0"
furo = ">=2024.8.6"
isort = ">=5.13.2"
mypy = ">=1.13.0"
pep8-naming = ">=0.14.1"
pre-commit = ">=4.0.1"
pre-commit-hooks = ">=5.0.0"
pytest = ">=8.3.4"
pyupgrade = ">=3.19.0"
safety = ">=3.2.11"
sphinx = ">=8.1.3"
sphinx-autobuild = ">=2024.10.3"
sphinx-click = ">=6.0.0"
typeguard = ">=4.4.1"
xdoctest = {extras = ["colors"], version = ">=1.2.0"}
myst-parser = {version = ">=4.0.0"}


[tool.poetry.scripts]
tradesignals-platform = "tradesignals.app:start"
start-pipeline = "pipelines.runner:main"


[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["tradesignals", "streaming", "pipelines", "core", "cli", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.flake8]
max-line-length = 120

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120

[tool.mypy]
strict = false
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
